use signed integer types for the floating size constraints

+++ include/config.h
@@ -151,10 +151,10 @@ struct Config {
     uint32_t floating_modifier;
 
     /** Maximum and minimum dimensions of a floating window */
-    uint16_t floating_maximum_width;
-    uint16_t floating_maximum_height;
-    uint16_t floating_minimum_width;
-    uint16_t floating_minimum_height;
+    int32_t floating_maximum_width;
+    int32_t floating_maximum_height;
+    int32_t floating_minimum_width;
+    int32_t floating_minimum_height;
 
     /* Color codes are stored here */
     struct config_client {
